DOM manipulation added
