+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "private": true,
   "scripts": {
-    "test": "jest --coverage",
+    "test": "jest --coverage --env=jsdom",
     "build": "webpack",
     "start": "webpack serve --open"
   },
